Reorganize modules; begin porting tests to use TContract
