+++ sdk/src/chain/mock_chain/mod.rs
@@ -1,22 +1,22 @@
+mod trait_impls;
+pub mod genesis_info;
+pub use trait_impls::*;
+
 use crate::chain::*;
-use crate::contract::generator::{
-    CellQuery, CellQueryAttribute, QueryProvider, QueryStatement, TransactionProvider,
-};
-use crate::contract::schema::{BytesConversion, JsonByteConversion, JsonConversion, MolConversion};
+
+
 use ckb_chain_spec::consensus::{Consensus, ConsensusBuilder};
 use ckb_error::Error as CKBError;
-use ckb_jsonrpc_types::TransactionView as JsonTransaction;
-use ckb_resource::BUNDLED;
+
 use ckb_script::{TransactionScriptsVerifier, TxVerifyEnv};
-use ckb_traits::{CellDataProvider, HeaderProvider};
-use ckb_types::core::{BlockBuilder, BlockView, HeaderBuilder, TransactionBuilder};
-use ckb_types::packed::{CellInput, CellOutputBuilder, ScriptOptBuilder};
+use ckb_types::core::Cycle;
+use ckb_types::packed::ScriptOptBuilder;
 use ckb_types::{
     bytes::Bytes,
     core::{
-        cell::{CellMeta, CellMetaBuilder, ResolvedTransaction},
+        cell::{CellMetaBuilder, ResolvedTransaction},
         hardfork::HardForkSwitch,
-        Capacity, Cycle, DepType, EpochExt, EpochNumberWithFraction, HeaderView, ScriptHashType,
+        Capacity, DepType, EpochExt, EpochNumberWithFraction, HeaderView, ScriptHashType,
         TransactionInfo, TransactionView,
     },
     packed::{Byte32, CellDep, CellOutput, OutPoint, Script},
@@ -24,11 +24,11 @@ use ckb_types::{
 use ckb_util::LinkedHashSet;
 use rand::{thread_rng, Rng};
 
-use ckb_always_success_script::ALWAYS_SUCCESS;
-use ckb_system_scripts::BUNDLED_CELL;
+
 
 use std::sync::{Arc, Mutex};
-use std::{cell::RefCell, collections::HashMap};
+use std::{collections::HashMap};
+
 const MAX_CYCLES: u64 = 500_0000;
 
 pub fn random_hash() -> Byte32 {
@@ -59,62 +59,6 @@ pub struct MockChain {
     messages: Arc<Mutex<Vec<Message>>>,
 }
 
-impl Default for MockChain {
-    fn default() -> Self {
-        let mut chain = Self {
-            cells: Default::default(),
-            outpoint_txs: Default::default(),
-            headers: Default::default(),
-            epoches: Default::default(),
-            cells_by_data_hash: Default::default(),
-            cells_by_lock_hash: Default::default(),
-            cells_by_type_hash: Default::default(),
-            genesis_info: None,
-            default_lock: None,
-            debug: Default::default(),
-            messages: Default::default(),
-        };
-
-        // Deploy always success script as default lock script
-        // This is required to deploy a random cell during the genesis event
-        let default_lock = chain.deploy_cell_with_data(Bytes::from(ALWAYS_SUCCESS.to_vec()));
-        chain.default_lock = Some(default_lock);
-
-        // Deploy system scripts to the chain
-
-        // Run genesis event on the mockchain
-        genesis_event(&mut chain);
-
-        
-
-        // Return chain
-        chain
-    }
-}
-
-impl PartialEq for MockChain {
-    // Simple equality check for testing purposes
-    // Curves around genesis info not implementing PartialEq
-    fn eq(&self, other: &Self) -> bool {
-        self.cells == other.cells &&
-        self.default_lock == other.default_lock &&
-        self.outpoint_txs == other.outpoint_txs &&
-        self.headers == other.headers &&
-        self.epoches == other.epoches &&
-        self.cells_by_data_hash == other.cells_by_data_hash &&
-        self.cells_by_lock_hash == other.cells_by_lock_hash &&
-        self.cells_by_type_hash == other.cells_by_type_hash &&  
-        self.debug == other.debug &&
-
-        // Compare GenesisInfo
-        self.genesis_info.as_ref().unwrap().sighash_data_hash() == other.genesis_info.as_ref().unwrap().sighash_data_hash() &&
-        self.genesis_info.as_ref().unwrap().sighash_type_hash() == other.genesis_info.as_ref().unwrap().sighash_type_hash() &&
-        self.genesis_info.as_ref().unwrap().multisig_data_hash() == other.genesis_info.as_ref().unwrap().multisig_data_hash() &&
-        self.genesis_info.as_ref().unwrap().multisig_type_hash() == other.genesis_info.as_ref().unwrap().multisig_type_hash() &&
-        self.genesis_info.as_ref().unwrap().dao_data_hash() == other.genesis_info.as_ref().unwrap().dao_data_hash() &&
-        self.genesis_info.as_ref().unwrap().dao_type_hash() == other.genesis_info.as_ref().unwrap().dao_type_hash()
-    }
-}
 
 impl MockChain {
     pub fn deploy_cell_with_data(&mut self, data: Bytes) -> OutPoint {
@@ -462,205 +406,6 @@ impl MockChain {
     }
 }
 
-impl CellDataProvider for MockChain {
-    // load Cell Data
-    fn load_cell_data(&self, cell: &CellMeta) -> Option<Bytes> {
-        cell.mem_cell_data
-            .as_ref()
-            .map(|data| Bytes::from(data.to_vec()))
-            .or_else(|| self.get_cell_data(&cell.out_point))
-    }
-
-    fn get_cell_data(&self, out_point: &OutPoint) -> Option<Bytes> {
-        self.cells
-            .get(out_point)
-            .map(|(_, data)| Bytes::from(data.to_vec()))
-    }
-
-    fn get_cell_data_hash(&self, out_point: &OutPoint) -> Option<Byte32> {
-        self.cells
-            .get(out_point)
-            .map(|(_, data)| CellOutput::calc_data_hash(data))
-    }
-}
-
-impl HeaderProvider for MockChain {
-    // load header
-    fn get_header(&self, block_hash: &Byte32) -> Option<HeaderView> {
-        self.headers.get(block_hash).cloned()
-    }
-}
-
-pub struct MockChainTxProvider {
-    pub chain: RefCell<MockChain>,
-}
-
-impl MockChainTxProvider {
-    pub fn new(chain: MockChain) -> Self {
-        Self {
-            chain: RefCell::new(chain),
-        }
-    }
-}
-
-impl TransactionProvider for MockChainTxProvider {
-    fn send_tx(&self, tx: JsonTransaction) -> Option<ckb_jsonrpc_types::Byte32> {
-        let mut chain = self.chain.borrow_mut();
-        let inner_tx = tx.inner;
-        let inner_tx = ckb_types::packed::Transaction::from(inner_tx);
-        let converted_tx_view = inner_tx.as_advanced_builder().build();
-        let tx = chain.complete_tx(converted_tx_view);
-        if let Ok(hash) = chain.receive_tx(&tx) {
-            let tx_hash: ckb_jsonrpc_types::Byte32 = hash.into();
-            Some(tx_hash)
-        } else {
-            None
-        }
-    }
-
-    fn verify_tx(&self, tx: JsonTransaction) -> bool {
-        let mut chain = self.chain.borrow_mut();
-        let inner_tx = tx.inner;
-        let inner_tx = ckb_types::packed::Transaction::from(inner_tx);
-        let converted_tx_view = inner_tx.as_advanced_builder().build();
-        let tx = chain.complete_tx(converted_tx_view);
-        println!(
-            "TX AFTER CHAIN COMPLETE {:#?}",
-            ckb_jsonrpc_types::TransactionView::from(tx.clone())
-        );
-        let result = chain.verify_tx(&tx, MAX_CYCLES);
-        match result {
-            Ok(_) => true,
-            Err(e) => {
-                println!("Error in tx verify: {:?}", e);
-                false
-            }
-        }
-    }
-}
-
-impl QueryProvider for MockChainTxProvider {
-    fn query_cell_meta(&self, query: CellQuery) -> Option<Vec<CellMeta>> {
-        if let Some(outpoints) = self.query(query) {
-            println!("OUTPOINTS TO CREATE CELL META: {:?}", outpoints);
-            Some(
-                outpoints
-                    .iter()
-                    .map(|outp| {
-                        let outp = ckb_types::packed::OutPoint::from(outp.clone());
-                        let cell_output = self.chain.borrow().get_cell(&outp).unwrap();
-                        CellMetaBuilder::from_cell_output(cell_output.0, cell_output.1)
-                            .out_point(outp)
-                            .build()
-                    })
-                    .collect(),
-            )
-        } else {
-            println!("NO OUTPOINTS TO RESOLVE IN QUERY CELL META");
-            None
-        }
-    }
-    fn query(&self, query: CellQuery) -> Option<Vec<ckb_jsonrpc_types::OutPoint>> {
-        let CellQuery { _query, _limit } = query;
-        println!("QUERY FROM QUERY PROVIDER: {:?}", _query);
-        match _query {
-            QueryStatement::Single(query_attr) => match query_attr {
-                CellQueryAttribute::LockHash(hash) => {
-                    let cells = self.chain.borrow().get_cells_by_lock_hash(hash.into());
-                    Some(
-                        cells
-                            .unwrap()
-                            .into_iter()
-                            .map(|outp| outp.into())
-                            .collect::<Vec<ckb_jsonrpc_types::OutPoint>>(),
-                    )
-                }
-                CellQueryAttribute::LockScript(script) => {
-                    let script = ckb_types::packed::Script::from(script);
-                    let cells = self
-                        .chain
-                        .borrow()
-                        .get_cells_by_lock_hash(script.calc_script_hash());
-                    Some(
-                        cells
-                            .unwrap()
-                            .into_iter()
-                            .map(|outp| outp.into())
-                            .collect::<Vec<ckb_jsonrpc_types::OutPoint>>(),
-                    )
-                }
-                CellQueryAttribute::TypeScript(script) => {
-                    let script = ckb_types::packed::Script::from(script);
-                    let cells = self
-                        .chain
-                        .borrow()
-                        .get_cells_by_type_hash(script.calc_script_hash());
-                    Some(
-                        cells
-                            .unwrap()
-                            .into_iter()
-                            .map(|outp| outp.into())
-                            .collect::<Vec<ckb_jsonrpc_types::OutPoint>>(),
-                    )
-                }
-                CellQueryAttribute::DataHash(hash) => Some(vec![self
-                    .chain
-                    .borrow()
-                    .get_cell_by_data_hash(&hash.into())
-                    .unwrap()
-                    .into()]),
-                _ => panic!("Capacity based queries currently unsupported!"),
-            },
-            _ => panic!("Compund queries currently unsupported!"),
-        }
-    }
-}
-
-impl Chain for MockChain {
-    type Inner = MockChainTxProvider;
-
-    fn inner(&self) -> Self::Inner {
-        MockChainTxProvider::new(self.clone())
-    }
-
-    fn deploy_cell(&mut self, cell: &Cell) -> ChainResult<OutPoint> {
-        let (outp, data): CellOutputWithData = cell.into();
-        Ok(self.deploy_cell_output(data, outp))
-    }
-
-    // Check how the genesis block is deployed on actual chains
-    fn genesis_info(&self) -> Option<GenesisInfo> {
-        self.genesis_info.clone()
-    }
-
-    fn set_genesis_info(&mut self, genesis_info: GenesisInfo) {
-        self.genesis_info = Some(genesis_info);
-    }
-
-    fn set_default_lock<A, D>(&mut self, lock: Contract<A, D>)
-    where
-        D: JsonByteConversion + MolConversion + BytesConversion + Clone + Default,
-        A: JsonByteConversion + MolConversion + BytesConversion + Clone + Default,
-    {
-        let (outp, data) = lock.as_code_cell();
-        let outpoint = self.deploy_cell_output(data, outp);
-        self.default_lock = Some(outpoint);
-    }
-
-    fn generate_cell_with_default_lock(&self, lock_args: crate::types::bytes::Bytes) -> Cell {
-        let script = self
-            .build_script(
-                &self.get_default_script_outpoint(),
-                lock_args.clone().into(),
-            )
-            .unwrap();
-        let mut cell = Cell::default();
-        cell.set_lock_script(script).unwrap();
-        cell.set_lock_args(lock_args).unwrap();
-        cell
-    }
-}
-
 // // Deploy system scripts from ckb-system-scripts bundled cell
 // fn genesis_event(chain: &mut MockChain) {
 //     todo!()
@@ -679,129 +424,10 @@ impl Chain for MockChain {
 //     Ok(outpoint)
 // }
 
-struct GenesisScripts {
-    secp256k1_data: Bytes,
-    secp256k1_blake160_sighash_all: Bytes,
-    secp256k1_blake160_multisig_all: Bytes,
-    dao: Bytes,
-}
-
-impl Default for GenesisScripts {
-    fn default() -> Self {
-        let bundle = &BUNDLED_CELL;
-        GenesisScripts {
-            secp256k1_data: Bytes::from(bundle.get("specs/cells/secp256k1_data").unwrap().to_vec()),
-            secp256k1_blake160_sighash_all: Bytes::from(
-                bundle
-                    .get("specs/cells/secp256k1_blake160_sighash_all")
-                    .unwrap()
-                    .to_vec(),
-            ),
-            secp256k1_blake160_multisig_all: Bytes::from(
-                bundle
-                    .get("specs/cells/secp256k1_blake160_multisig_all")
-                    .unwrap()
-                    .to_vec(),
-            ),
-            dao: Bytes::from(bundle.get("specs/cells/dao").unwrap().to_vec()),
-        }
-    }
-}
-
-// Deploy every system script from a genesis script to a MockChain return GenesisInfo
-fn genesis_event(
-    chain: &mut MockChain,
-) -> GenesisInfo {
-    // Does nothing if the genesis scripts are already deployed
-    match &chain.genesis_info {
-        // Should return the existing genesis info
-        Some(genesis_info) => {
-            genesis_info.clone()
-        }
-
-        // Run genesis event
-        // Deploy genesis scripts, create a block, set the genesis info and return it
-        None => {
-            // Deploy scripts
-            deploy_genesis_scripts(chain, None);
-            // Create genesis block
-            let block = genesis_block_from_chain(chain);
-            // Insert block header
-            chain.insert_header(block.clone().header());
-            // Return genesis info
-            GenesisInfo::from_block(&block).expect("Failed to create genesis info from genesis block")
-        }
-    }
-}
-
-fn deploy_genesis_scripts(
-    chain: &mut MockChain,
-    scripts: Option<GenesisScripts>
-) {
-    // Generate default scripts if no scripts were passed
-    let scripts = scripts.unwrap_or_default();
-
-    // Deploy every script to the chain
-    chain.deploy_cell_with_data(scripts.secp256k1_data.clone());
-    chain.deploy_cell_with_data(scripts.secp256k1_blake160_sighash_all.clone());
-    chain.deploy_cell_with_data(scripts.secp256k1_blake160_multisig_all.clone());
-    chain.deploy_cell_with_data(scripts.dao.clone());
-}
-
-
-
-fn genesis_block_from_chain(chain: &mut MockChain) -> BlockView {
-    let block: BlockBuilder = BlockBuilder::default();
-
-    let tx = TransactionBuilder::default();
-
-    let secp256k1_data_code_hash_bytes =
-        Byte32::from_slice(&ckb_system_scripts::CODE_HASH_SECP256K1_DATA).unwrap();
-    let secp256k1_data_outpoint = chain
-        .get_cell_by_data_hash(&secp256k1_data_code_hash_bytes)
-        .unwrap();
-    let secp256k1_data = chain.get_cell(&secp256k1_data_outpoint).unwrap();
-    let tx = tx.output(secp256k1_data.0.clone());
-    let tx = tx.output_data(secp256k1_data.1.clone().pack());
-
-    let blake_sighash_all_code_hash_bytes =
-        Byte32::from_slice(&ckb_system_scripts::CODE_HASH_SECP256K1_BLAKE160_SIGHASH_ALL).unwrap();
-    let blake_sighash_all_outpoint = chain
-        .get_cell_by_data_hash(&blake_sighash_all_code_hash_bytes)
-        .unwrap();
-    let blake_sighash_all = chain.get_cell(&blake_sighash_all_outpoint).unwrap();
-    let tx = tx.output(blake_sighash_all.0.clone());
-    let tx = tx.output_data(blake_sighash_all.1.clone().pack());
-
-    let dao_code_hash_bytes = Byte32::from_slice(&ckb_system_scripts::CODE_HASH_DAO).unwrap();
-    let dao_outpoint = chain.get_cell_by_data_hash(&dao_code_hash_bytes).unwrap();
-    let dao = chain.get_cell(&dao_outpoint).unwrap();
-    let tx = tx.output(dao.0.clone());
-    let tx = tx.output_data(dao.1.clone().pack());
-
-    // Some cell without data or scripts to complete the genesis block and respect the script order
-    let random_cell_outpoint = chain.deploy_random_cell_with_default_lock(100000, None);
-    let random_cell = chain.get_cell(&random_cell_outpoint).unwrap();
-    let tx = tx.output(random_cell.0.clone());
-    let tx = tx.output_data(random_cell.1.clone().pack());
-
-    let blake_multisig_code_hash_bytes =
-        Byte32::from_slice(&ckb_system_scripts::CODE_HASH_SECP256K1_BLAKE160_MULTISIG_ALL).unwrap();
-    let blake_multisig_outpoint = chain
-        .get_cell_by_data_hash(&blake_multisig_code_hash_bytes)
-        .unwrap();
-    let blake_multisig = chain.get_cell(&blake_multisig_outpoint).unwrap();
-    let tx = tx.output(blake_multisig.0.clone());
-    let tx = tx.output_data(blake_multisig.1.clone().pack());
-
-    let block = block.transaction(tx.build());
-
-    block.build()
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use genesis_info::*;
 
     fn mockchain_setup() -> (mock_chain::MockChain, ckb_types::core::BlockView) {
         // Create a new mockchain
@@ -841,7 +467,7 @@ mod tests {
 
     #[test]
     fn genesis_info_from_genesis_block_returns_ok() {
-        let (chain, genesis_block) = mockchain_setup();
+        let (_chain, genesis_block) = mockchain_setup();
 
         let genesis_info = GenesisInfo::from_block(&genesis_block);
 
@@ -850,7 +476,7 @@ mod tests {
 
     #[test]
     fn test_genesis_block_has_dao_cell() {
-        let (mut chain, genesis_block) = mockchain_setup();
+        let (chain, genesis_block) = mockchain_setup();
 
         // Get the cell by hash
         let dao_code_hash_bytes = Byte32::from_slice(&ckb_system_scripts::CODE_HASH_DAO).unwrap();
@@ -871,7 +497,7 @@ mod tests {
 
     #[test]
     fn test_genesis_block_has_secp_multisig_cell() {
-        let (mut chain, genesis_block) = mockchain_setup();
+        let (chain, genesis_block) = mockchain_setup();
 
         // Get the cell by hash
         let multisig_code_hash_bytes =
@@ -930,10 +556,10 @@ mod tests {
         let mut chain = MockChain::default();
 
         // Create default genesis scripts
-        let genesis_scripts = GenesisScripts::default();
+        let _genesis_scripts = GenesisScripts::default();
 
         // Run genesis event on the mockchain with the scripts
-        let scripts = genesis_event(&mut chain);
+        let _scripts = genesis_event(&mut chain);
 
         // Generate genesis block
         let genesis_block = genesis_block_from_chain(&mut chain);
