+++ sdk/src/contract/builtins/sudt/mod.rs
@@ -1,7 +1,7 @@
 use std::prelude::v1::*;
 
 use crate::ckb_types::packed::{Byte32, Uint128};
-use crate::contract::Contract;
+use crate::contract::{Contract, TContract};
 
 use crate::contract::schema::SchemaPrimitiveType;
 
@@ -15,3 +15,5 @@ pub type OwnerLockHash = SchemaPrimitiveType<[u8; 32], Byte32>;
 pub type SudtAmount = SchemaPrimitiveType<u128, Uint128>;
 
 pub type SudtContract = Contract<OwnerLockHash, SudtAmount>;
+
+pub type SudtTrampolineContract = TContract<OwnerLockHash, SudtAmount>;
